refactor(userspace): extract binding/queue control+status to protocol_binding.go
Ninth code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the forwarding/queue/binding control-request types, the
per-queue and per-binding status payloads, and the inject-packet request
into a cohesive sibling protocol_binding.go in the same userspace
package.

Moved verbatim (no rename, no logic or json-tag change):
ForwardingControlRequest, QueueControlRequest, BindingControlRequest,
QueueStatus, BindingStatus, BindingCountersSnapshot, InjectPacketRequest.

The time import migrates with BindingStatus/BindingCountersSnapshot's
time.Time fields. Behavior byte-identical; json tags preserved for the
Rust userspace-dp control wire. Build passes; gofmt clean.
